refactor(#9): Moved types file to services folder
